Move FeatureFactory files to di/ package
Factory files are now in di/ which is already ignored by codecov.
Removed the separate *FeatureFactory.kt rule.
