+++ codecov.yml
@@ -11,7 +11,6 @@ coverage:
 ignore:
   - "**/di/**"
   - "**/view/**"
-  - "**/*FeatureFactory.kt"
 
 comment:
   layout: "reach, diff, flags, files"
