+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/di/HabitsFeatureFactory.kt
@@ -1,7 +1,12 @@
-package space.be1ski.vibits.shared.feature.habits.presentation
+package space.be1ski.vibits.shared.feature.habits.di
 
 import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsAction
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsEffect
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsState
+import space.be1ski.vibits.shared.feature.habits.presentation.habitsReducer
 import space.be1ski.vibits.shared.feature.memos.domain.repository.MemosRepository
 
 /**
